More Code Moved from Admin
