+++ Database/Seeders/AuthDatabaseSeeder.php
@@ -16,6 +16,6 @@ class AuthDatabaseSeeder extends Seeder
     {
         Model::unguard();
 
-        // $this->call("OthersTableSeeder");
+        $this->call(UserRegisterSeeder::class);
     }
 }
