+++ Models/Login.php
@@ -1,0 +1,76 @@
+<?php
+
+namespace Modules\Auth\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\SoftDeletes;
+use Kyslik\ColumnSortable\Sortable;
+use Modules\Core\Models\User;
+use OwenIt\Auditing\Auditable as AuditableTrait;
+use OwenIt\Auditing\Contracts\Auditable;
+
+class Login extends Model implements Auditable
+{
+    use AuditableTrait, SoftDeletes, Sortable;
+
+    /**
+     * The attributes that are mass assignable.
+     * 'enabled' => to handle status,
+     * ['created_by', 'updated_by', 'deleted_by'] => for audit
+     *
+     * @var array
+     */
+    protected $fillable = ['name', 'guard_name', 'remarks', 'enabled'];
+
+
+    /**
+     * The attributes that should be hidden for arrays.
+     *
+     * @var array
+     */
+    protected $hidden = [];
+
+    /**
+     * The attributes that should be cast to native types.
+     *
+     * @var array
+     */
+    protected $casts = [];
+
+    /**
+     * The model's default values for attributes when new instance created.
+     *
+     * @var array
+     */
+    protected $attributes = [
+        'guard_name' => 'web',
+        'enabled' => 'yes'
+    ];
+
+    /************************ Audit Relations ************************/
+
+    /**
+     * @return BelongsTo
+     */
+    public function createdBy(): BelongsTo
+    {
+        return $this->belongsTo(User::class, 'created_by');
+    }
+
+    /**
+     * @return BelongsTo
+     */
+    public function updatedBy(): BelongsTo
+    {
+        return $this->belongsTo(User::class, 'updated_by');
+    }
+
+    /**
+     * @return BelongsTo
+     */
+    public function deletedBy(): BelongsTo
+    {
+        return $this->belongsTo(User::class, 'deleted_by');
+    }
+}
