+++ Repositories/Eloquent/LoginRepository.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Modules\Auth\Repositories\Eloquent;
+
+class LoginRepository
+{
+
+}
